pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screen-vision"
version = "0.1.0"
description = "MCP server that gives Claude Code the ability to see your screen"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
full = [
    "pytesseract>=0.3.10",
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "opencv-python>=4.8.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
screen-vision-mcp = "screen_vision.server:main"

[tool.setuptools.packages.find]
where = ["src"]
